Move Cookbook controller outside Exporter
