+++ app/controllers/__init__.py
@@ -3,6 +3,7 @@ from .daily_plans import DailyPlanView
 from .daily_plan_tasks import DailyPlanTaskView
 from .daily_plan_recipes import DailyPlanRecipeView
 from .event_exporter import EventExporterView
+from .exports.event_cookbook_exporter import EventCookbookExporterView
 
 from .events.events import EventView
 from .events.share_events import ShareEventView
@@ -42,6 +43,7 @@ __all__ = [
     "DailyPlanRecipeView",
     "EventExporterView",
     "EventView",
+    "EventCookbookExporterView",
     "ShareEventView",
     "PublishedEventView",
     "ErrorView",
@@ -81,6 +83,7 @@ def register_all_controllers(application):
     PublishedEventView.register(application, base_class=HelperFlaskView)
 
     EventExporterView.register(application, base_class=HelperFlaskView)
+    EventCookbookExporterView.register(application, base_class=HelperFlaskView)
 
     ErrorView.register(application)
     SupportView.register(application)
