+++ app/controllers/event_exporter.py
@@ -1,5 +1,4 @@
 from flask import redirect, url_for
-from flask import render_template as template
 from flask_security import login_required
 from flask_weasyprint import render_pdf, HTML
 
@@ -18,23 +17,16 @@ class EventExporterView(HelperFlaskView):
         self.event = Event.load(event_id)
         self.validate_show(self.event)
 
+    def _show_list(self, event_id, is_print=False):
+        self._set_values()
+
+        return self.template(template_name="shopping_list", print=is_print)
+
     def show_list(self, event_id):
         return self._show_list(event_id)
 
-    def show_table(self, event_id):
-        return self._show_table(event_id)
-
     def show_list_pdf(self, event_id):
         string = self._show_list(event_id, is_print=True)
-
-        return render_pdf(
-            HTML(string=string),
-            download_filename=f"{self.event.slugified_name}--nakupni-seznam.pdf",
-            automatic_download=False,
-        )
-
-    def show_table_pdf(self, event_id):
-        string = self._show_table(event_id, is_print=True)
 
         return render_pdf(
             HTML(string=string),
@@ -49,6 +41,20 @@ class EventExporterView(HelperFlaskView):
             download_filename=f"{self.event.slugified_name}--nakupni-seznam.pdf",
         )
 
+    # Shopping list export
+
+    def show_table(self, event_id):
+        return self._show_table(event_id)
+
+    def show_table_pdf(self, event_id):
+        string = self._show_table(event_id, is_print=True)
+
+        return render_pdf(
+            HTML(string=string),
+            download_filename=f"{self.event.slugified_name}--nakupni-seznam.pdf",
+            automatic_download=False,
+        )
+
     def download_table_pdf(self, event_id):
         string = self._show_list(event_id, is_print=True)
         return render_pdf(
@@ -56,7 +62,10 @@ class EventExporterView(HelperFlaskView):
             download_filename=f"{self.event.slugified_name}--nakupni-seznam.pdf",
         )
 
-    # Shopping list export
+    def _show_table(self, event_id, is_print=False):
+        self._set_values()
+
+        return self.template(template_name="shopping_table", print=is_print)
 
     def show_shopping_list(self, event_id):
         return redirect(url_for("EventExporterView:show_table", event_id=event_id))
@@ -111,41 +120,6 @@ class EventExporterView(HelperFlaskView):
             download_filename=f"{self.event.slugified_name}--recepty.pdf",
         )
 
-    # Cookbook export
-    def _cookbook(self, is_print=False):
-        partial_templates = []
-        for daily_plan in self.event.active_daily_plans:
-            for daily_recipe in daily_plan.daily_recipes:
-                recipe = daily_recipe.recipe
-                if recipe.is_shopping:
-                    continue
-                recipe.reload()
-                recipe.portion_count = daily_recipe.portion_count
-                recipe_template = template(
-                    "recipes/_show_simple.html.j2", recipe=recipe, print=is_print
-                )
-                partial_templates.append(recipe_template)
-
-        self.recipes_html = "".join(partial_templates)
-
-        return self.template(template_name="cookbook", print=is_print)
-
-    def show_cookbook(self, event_id, **kwargs):
-        return self._cookbook()
-
-    def show_cookbook_pdf(self, event_id):
-        return render_pdf(
-            HTML(string=self._cookbook(is_print=True)),
-            download_filename=f"{self.event.slugified_name}--kucharka.pdf",
-            automatic_download=False,
-        )
-
-    def download_cookbook_pdf(self, event_id):
-        return render_pdf(
-            HTML(string=self._cookbook(is_print=True)),
-            download_filename=f"{self.event.slugified_name}--kucharka.pdf",
-        )
-
     # INTERNAL
 
     def _set_values(self):
@@ -166,13 +140,3 @@ class EventExporterView(HelperFlaskView):
             shopping.recipe_ingredient_amounts = (
                 self.shopping_list_constructor.get_amounts_for_shopping(shopping)
             )
-
-    def _show_list(self, event_id, is_print=False):
-        self._set_values()
-
-        return self.template(template_name="shopping_list", print=is_print)
-
-    def _show_table(self, event_id, is_print=False):
-        self._set_values()
-
-        return self.template(template_name="shopping_table", print=is_print)
