+++ app/templates/events/_export.html.j2
@@ -13,7 +13,7 @@
 			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventExporterView:show_recipe_list', event_id=event.id) }}"> {{ icon("food") }} jídelníček </a>
 		</div>
 		<div class="col-12 col-sm-3">
-			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventExporterView:show_cookbook', event_id=event.id) }}"> {{ icon("cookbook") }} kuchařka </a>
+			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventCookbookExporterView:show', event_id=event.id) }}"> {{ icon("cookbook") }} kuchařka </a>
 		</div>
 		</div>
 	</div>
@@ -32,7 +32,7 @@
 			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventExporterView:show_recipe_list_pdf', event_id=event.id) }}" target="_blank"> {{ icon("pdf") }} jídelníček  </a>
 		</div>
 		<div class="col-12 col-sm-3">
-			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventExporterView:show_cookbook_pdf', event_id=event.id) }}" target="_blank"> {{ icon("pdf") }} kuchařka </a>
+			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventCookbookExporterView:pdf', event_id=event.id) }}" target="_blank"> {{ icon("pdf") }} kuchařka </a>
 		</div>
 		</div>
 	</div>
@@ -51,7 +51,7 @@
 			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventExporterView:download_recipe_list_pdf', event_id=event.id) }}" target="_blank"> {{ icon("pdf") }} jídelníček  </a>
 		</div>
 		<div class="col-12 col-sm-3">
-			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventExporterView:download_cookbook_pdf', event_id=event.id) }}" target="_blank"> {{ icon("pdf") }} kuchařka </a>
+			<a data-turbo="false" class="btn bg-color-secondary-action color-white me-2 mt-1 mt-lg-0 w-100" href="{{ url_for('EventCookbookExporterView:download', event_id=event.id) }}" target="_blank"> {{ icon("pdf") }} kuchařka </a>
 		</div>
 		</div>
 	</div>
